feat(comparesql): Implement comprehensive SQL comparison strategies
- Added strategies for handling various stages of SQL comparison, including:
  - AskFirstSQLMethod
  - NeedFirstNaturalLanguage
  - NeedFirstUserSQL
  - ConfirmFirstSQL
  - AskSecondSQLMethod
  - NeedSecondNaturalLanguage
  - NeedSecondUserSQL
  - ConfirmSecondSQL
  - AskAutoMatch
  - WaitingMapTable
  - AskReportingType
  - AskCompareSchema
  - AskCompareTableName
  - ExecuteCompareSQL
  - AskCompareJobName

- Each strategy includes logic for user input handling, SQL generation, and error management.
- Introduced JSON handling for mapping tables and reporting types.
- Enhanced logging for better traceability of user interactions and errors.
